Small configuration changes
Changed the general requirement from the JUnit API to the JUnit Engine, and added the Surefire plugin to the POM. Economy has been configured to run the tests through maven.

Properties module descriptor now requires java.base module. This is not that important as it seemed to compile and work without it, but added just for correctness