Released Scoreboard
Scoreboard module was lost at some point before I started using git
When I recreated it I didn't know if it was feature complete.
While revisiting the project that it was built for, I have confirmed
that all necessary features exist. Thus, a full release is welcome.
Future plans for the scoreboard exist