The Bible Update
This commit insures that all committed modules have 100% javadoc coverage, it insures that in the future the intention behind methods will be clear. It also provides small optimizations across the board, updating old snippets into better, more modern ones (eg. a chain of "if else" statements being converted to a "switch" statement).

Overall, it should be a big commit but a very needed one