Work on IerisFX, Config and Assets
IerisFX was an abstraction of some leftover code. As such, it never was much in a working state. This commit brings it to a working state. Much work is left to be done, but this is a start. The changes to Config and Assets is still relevant to those modules, but is there to support configuration of IerisFX.