Make ierislog a top level module
The project has grown so much, along the changes coming now, it no longer makes sense to hide it under utils. It should, like ierisfx, be front and center in this library.

For now this means deleting all the previous work. I want to rewrite it from scratch to make a cleaner, more functional logger. I have kept a local copy of it to reference for the rewrite, but otherwise, the old code ought to go