Revamp IerisFX class
Heavily improved documentation on this class
Added a child Application inheritor instead of inheriting itself to better control user access. The visibility of said class is untested as of this point.
Replaced Scripts interfaces for Runnables. They have the same functionality and don't provide the false pretense that an array of strings "args" will be given to the method as a parameter
Added logging